BFWAnimationView: changed animation, start, end and cycles to double type

+++ BFWDrawView/Modules/Draw/View/BFWAnimationView.h
@@ -12,11 +12,11 @@ IB_DESIGNABLE
 
 @interface BFWAnimationView : BFWDrawView
 
-@property (nonatomic, assign) IBInspectable CGFloat animation; // fraction 0.0 to 1.0. Set internally but exposed for storyboard preview
-@property (nonatomic, assign) IBInspectable CGFloat start; // fraction 0.0 to 1.0. Start of animation.
-@property (nonatomic, assign) IBInspectable CGFloat end; // fraction 0.0 to 1.0. End of animation.
+@property (nonatomic, assign) IBInspectable double animation; // fraction 0.0 to 1.0. Set internally but exposed for storyboard preview
+@property (nonatomic, assign) IBInspectable double start; // fraction 0.0 to 1.0. Start of animation.
+@property (nonatomic, assign) IBInspectable double end; // fraction 0.0 to 1.0. End of animation.
 @property (nonatomic, assign) IBInspectable double duration; // default = 3 seconds
-@property (nonatomic, assign) IBInspectable NSUInteger cycles; // default 0 = infinite repetitions
+@property (nonatomic, assign) IBInspectable double cycles; // default 0 = infinite repetitions
 @property (nonatomic, assign) IBInspectable BOOL paused;
 
 @property (nonatomic, assign) CGFloat framesPerSecond;
